fix(schema): sync cloud with desktop app data model
Two drifts the desktop app had moved past the cloud/repo schema:
- products.quantity/reorder_level and order_items.quantity were bigint but
  the app stores fractional quantities (oil sold by the litre); widen to
  numeric(14,3) to match purchase_order_items/stock_movements.
- invoice_docs missing show_bank + advance_applied (show_logo already in
  invoice-missing-columns.sql); add them.

Applied live to cloud 2026-07-11 + migration for fresh deploys.

+++ supabase/schema.sql
@@ -81,8 +81,8 @@ create table if not exists products (
   category text,
   unit_price numeric(14,2) not null default 0,
   cost_price numeric(14,2) not null default 0,
-  quantity bigint not null default 0,
-  reorder_level bigint not null default 0,
+  quantity numeric(14,3) not null default 0,
+  reorder_level numeric(14,3) not null default 0,
   created_at timestamptz not null default now(),
   updated_at timestamptz not null default now()
 );
@@ -103,7 +103,7 @@ create table if not exists order_items (
   user_id uuid not null default auth.uid() references auth.users(id) on delete cascade,
   order_id bigint references orders(id) on delete cascade,
   product_id bigint,
-  quantity bigint not null default 1,
+  quantity numeric(14,3) not null default 1,
   unit_price numeric(14,2) not null default 0,
   created_at timestamptz not null default now(),
   updated_at timestamptz not null default now()
